requirements.txt
streamlit>=1.0.0
pandas>=1.0.0
PyMuPDF>=1.18.0
Pillow>=8.0.0